breeze-proxy/requirements.txt

flask
flask-cors
breeze_connect
google-cloud-secret-manager
requests
gunicorn
